requirements.txt
uvicorn~=0.30.6
fastapi~=0.112.2
pymongo~=4.8.0
pydantic~=2.8.2
python-dotenv